Add support for 32-bit types to transpose_dest.
As suggested by @corsix, this performs a transpose on a 16x16 face of 32-bit
values by splitting into two faces of 16-bit values (the low 16 bits and high
16 bits), transposing each face separately using `TRNSPSRCB`, and then
recombining into 32-bit values.

A neat way to achieve this (also suggested by @corsix!) is to use the
`dest_32b_lo` instruction mode flag on `MOVD2B`/`MOVB2D`.  This allows the low
bits to be copied to SrCB, transposed, and copied back.  Followed by the same
operation on the high bits.

Further optimisation would involve using the full 32 slots in the replay
buffer, rather than overwriting slots 16 to 32 as I'm doing currently.

This was just the most straightforward way to get a working proof of concept.