docs: add link to deepwiki (#372)
### Ticket
None

### Problem description
[Deepwiki](https://deepwiki.com/tenstorrent/tt-llk) is already indexing
our repository, but adding the badge will make it index it regularly. It
generates nice documentation and combines the code, and has integrated
AI bot who can answer questions for you. Could be useful for some new
developers. tt-metal is already using it, and it seems quite nice in
their case.

### What's changed
New badge is added to the README.

### Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update