Add num_faces argument to the packer functions and enable 32-bit mode ZEROACC in the datacopy.h
Packer functions now use the desired num_faces argument instead of using the default value 4. The functions vary a bit between BH and WH so this is taken into account using the ifdefs. The _llk_math_eltwise_unary_datacopy_ now includes a ZEROACC with 32-bit mode enabled, used when Float32 is used in the test.